Move out the /ingests tests into a separate file
